Separate build dirs for CUDA and CPU versions, move FM to special dir and extract common function to separated file
