+++ rebuild_and_run_cuda.sh
@@ -1,5 +1,4 @@
 #!/bin/sh
-mkdir build || rm ./build -r -f
-cmake . -B build -D USE_CUDA=YES
-make -C build -j8
+cmake . -B cuda_build -D USE_CUDA=YES
+make -C cuda_build -j8
 ./build/slam-indoor-code $1
