+++ src/legacy/mainCycle.cpp
@@ -9,7 +9,7 @@
 #include "../cameraCalibration.h"
 #include "../cameraTransition.h"
 #include "../fastExtractor.h"
-#include "../featureMatching.h"
+#include "../featureMatching/featureMatching.h"
 #include "../IOmisc.h"
 #include "../bundleAdjustment.h"
