+++ src/legacy/mainCycleInterNotWork.cpp
@@ -5,7 +5,7 @@
 #include "../cameraCalibration.h"
 #include "../cameraTransition.h"
 #include "../fastExtractor.h"
-#include "../featureMatching.h"
+#include "../featureMatching/featureMatching.h"
 #include "../IOmisc.h"
 
 #include "../config/config.h"
